pyproject.toml
[tool.poetry]
name = "teste-outsera"
version = "0.1.0"
package-mode = false
description = ""
authors = ["manoelmsnsi <manoelmsnsi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
flask-cors = "^5.0.1"
python-dotenv = "^1.0.1"
flask-sqlalchemy = "^3.1.1"
pydantic = "^2.10.6"
pytest = "^8.3.4"
flask-testing = "^0.8.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
